mint forknet added to pipeline
