+++ scripts/actions/deployContracts.js
@@ -5,7 +5,7 @@ const deployContracts = async (contractFactories, instances = []) => {
   const deployedInstance = await contractFactories[0].factory.deploy();
   instances[contractFactories[0].instance] = deployedInstance;
   console.log(
-    `${contractFactories[0].instance} address is: , ${deployedInstance.address}`
+    `${contractFactories[0].instance} address is: ${deployedInstance.address}`
   );
   return await deployContracts(
     contractFactories.slice(1, contractFactories.length),
