+++ scripts/actions/mintForknet.js
@@ -1,0 +1,75 @@
+const hre = require("hardhat");
+const ethers = hre.ethers;
+const { formatUnits, parseUnits } = ethers.utils;
+const { address } = require("../utils/helpers");
+
+const wbtc_address = "0x2260fac5e5542a773aa44fbcfedf7c193bc2c599";
+const cvx_address = "0x4e3fbd56cd56c3e72c1403e103b45db9da5b9d2b";
+
+const wbtc_minter_address = "0xCA06411bd7a7296d7dbdd0050DFc846E95fEBEB7"; // owner address of wbtc
+const cvx_minter_address = "0xF403C135812408BFbE8713b5A23a04b3D48AAE31"; // operator address of cvx
+
+const erc20_mintable_abi = ["function mint(address, uint256)"];
+
+const mintForknet = async ({ mintTo, ERC20Upgradeable }) => {
+  /// === mint wbtc and cvx to signers[0]
+  // impersonate the token owner
+  await hre.network.provider.request({
+    method: "hardhat_impersonateAccount",
+    params: [wbtc_minter_address],
+  });
+  await hre.network.provider.request({
+    method: "hardhat_impersonateAccount",
+    params: [cvx_minter_address],
+  });
+
+  // send some balance for the gas
+  await hre.network.provider.send("hardhat_setBalance", [
+    wbtc_minter_address,
+    "0x1000000000000000000",
+  ]);
+  await hre.network.provider.send("hardhat_setBalance", [
+    cvx_minter_address,
+    "0x1000000000000000000",
+  ]);
+
+  // get the signer
+  const wbtc_minter = await ethers.getSigner(wbtc_minter_address);
+  const cvx_minter = await ethers.getSigner(cvx_minter_address);
+
+  // connect the token contract to signers
+  const wbtcMintable = new ethers.Contract(
+    wbtc_address,
+    erc20_mintable_abi,
+    wbtc_minter
+  );
+
+  const cvxMintable = new ethers.Contract(
+    cvx_address,
+    erc20_mintable_abi,
+    cvx_minter
+  );
+
+  // mint some tokens to signers[0]
+  const user = mintTo;
+  const txWbtcMint = await wbtcMintable.mint(
+    address(user),
+    parseUnits("100", 8) // 100 btc
+  );
+  await txWbtcMint.wait();
+  await cvxMintable.mint(
+    address(user),
+    parseUnits("100000", 18) // 100000 cvx
+  );
+
+  const wbtc = ERC20Upgradeable.attach(wbtc_address); //
+  const cvx = ERC20Upgradeable.attach(cvx_address); //
+
+  // check the balance
+  const balance_wbtc = await wbtc.balanceOf(address(user));
+  console.log(`wbtc balance of signers[0]: ${formatUnits(balance_wbtc, 8)}`);
+  const balance_cvx = await cvx.balanceOf(address(user));
+  console.log(`cvx balance of signers[0]: ${formatUnits(balance_cvx, 18)}`);
+};
+
+module.exports = mintForknet;
